Add arg getters and refactor API to encapsulate configuration object
Made it so the libary user only has to instantiate the parser, add arguments, parse command line arges (todo), and get the arg values.

Added a function for getting arg values for each type of supported value. This way the logic on the user side is simplified compared to if they had to receive a variant.